requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django
django-cors-headers==3.2.1
django-elasticsearch-dsl==7.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt
drf-yasg==1.17.1
elasticsearch==7.5.1
elasticsearch-dsl==7.1.0
gunicorn==20.0.4
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.1
psycopg2==2.8.4
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2018.9
requests==2.22.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
whitenoise==5.0.1
pillow==7.1.1
